% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_BAM.R
\name{quality_control_BAM}
\alias{quality_control_BAM}
\title{Filters a BAM File Based on Quality Control Metrics}
\usage{
quality_control_BAM(
  bamPath,
  min_length = 50,
  min_depth = 10,
  outputBAM = "filtered_output.bam"
)
}
\arguments{
\item{bamPath}{A character string specifying the path to the input BAM file.}

\item{min_length}{An integer indicating the minimum read length to retain.
Reads shorter than this length will be excluded. Default is 50.}

\item{min_depth}{An integer specifying the minimum mapping quality (depth) required for each read.
Reads with a depth lower than this value will be excluded. Default is 10.}

\item{outputBAM}{A character string for the name of the output filtered BAM file. Default is "filtered_output.bam".}
}
\value{
A filtered BAM file that contains only high-quality reads meeting the specified criteria.
This output BAM file can be used for further analyses, such as variant calling and error rate calculation.
}
\description{
This function filters a BAM file based on specified quality control criteria, such as
minimum read length and minimum mapping quality. It excludes reads that do not meet
these requirements, ensuring that only high-quality reads are retained for downstream
analyses, such as variant calling.
}
\examples{
# Example: Filter a BAM file based on minimum length and depth requirements, replace it with local path.
filteredResult <- quality_control_BAM(
  bamPath = "C:/Users/rjay1/Desktop/BCB410/LegitXMut/inst/extdata/aligned_output.bam",
  min_length = 50,
  min_depth = 10,
  outputBAM = "C:/Users/rjay1/Desktop/BCB410/LegitXMut/inst/extdata/filtered_output.bam"
)

}
\references{
Morgan M, Pagès H, Obenchain V, Hayden N (2024). Rsamtools: Binary
alignment (BAM), FASTA, variant call (BCF), and tabix file import. R package
version 2.20.0, https://bioconductor.org/packages/Rsamtools.

OpenAI. ChatGPT: Assistance with R function development for bioinformatics applications,
"Assignment". https://chat.openai.com (2023, accessed 5 November 2024).
}
