% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLegitXMut.R
\name{runLegitXMut}
\alias{runLegitXMut}
\title{Launch Shiny App for LegitXMut}
\usage{
runLegitXMut()
}
\value{
No return value; launches a Shiny app in the browser.
}
\description{
A function that launches the Shiny app for the LegitXMut package.
The app provides a user-friendly interactive interface for alignment and visulization
using functions from the LegitXMut package.
}
\examples{
\dontrun{
LegitXMut::runLegitXMut()
}

}
\references{
Chang, W., J. Cheng, J. Allaire, C. Sievert, B. Schloerke, Y. Xie, J. Allen, J. McPherson, A. Dipert, B. Borges (2024).
shiny: Web Application Framework for R. R package version 1.9.1, https://CRAN.R-project.org/package=shiny

Silva, Anjali. TestingPackage. GitHub, https://github.com/anjalisilva/TestingPackage. Accessed 5 Nov. 2024. \href{https://shiny.rstudio.com/}{Link}
}
