% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vcf.R
\name{update_vcf}
\alias{update_vcf}
\title{Update Chromosome Names to VCF File Based on Reference FASTA If Performing Single Chromosome Anlysis}
\usage{
update_vcf(fastaPath, vcfPath, outputVcfPath)
}
\arguments{
\item{fastaPath}{A character string specifying the path to the reference FASTA file.}

\item{vcfPath}{A character string specifying the path to the input VCF file that needs chromosome name updating.}

\item{outputVcfPath}{A character string specifying the path where the updated VCF file will be saved.}
}
\value{
NULL. A message confirms that the VCF file has been updated and saved.
}
\description{
This function reads a reference FASTA file to extract chromosome mappings and updates
the chromosome names in a VCF file accordingly. The updated VCF file is saved to a specified output path.
}
\examples{
# Example: Update chromosome names in a VCF file using a reference FASTA file/FNA file
# Replace with actual test file path
fastaPath <- system.file("extdata", "yeast.fna", package = "LegitXMut")
vcfPath <- system.file("extdata", "aligned_output.bam.indel.vcf", package = "LegitXMut")
outputVcfPath <- file.path(system.file("extdata", package = "LegitXMut"), "updated.vcf")

# Update the VCF file with specified inputs
updatingvcf <- update_vcf(
  fastaPath = fastaPath,
  vcfPath = vcfPath,
  outputVcfPath = outputVcfPath
)

}
\references{
Morgan M, Pages H, Obenchain V, and Hayden N (2020).
“VariantAnnotation: annotation of genetic variants.”
Bioconductor. doi:10.18129/B9.bioc.VariantAnnotation. Available at: https://bioconductor.org/packages/release/bioc/html/VariantAnnotation.html.

Wickham, H. (2019). stringr: Simple, Consistent Wrappers for Common String Operations.
R package version 1.4.0. Available at: https://CRAN.R-project.org/package=stringr

Pagès, H., Aboyoun, P., DebRoy, S., and Lawrence, M. (2023).
GenomeInfoDb: Utilities for manipulating chromosome names, including modifying the names to follow a particular convention.
R package version 1.36.0, Bioconductor, https://bioconductor.org/packages/GenomeInfoDb.

OpenAI. ChatGPT: Assistance with R function development for bioinformatics applications,
"Assignment". https://chat.openai.com (2023, accessed 5 November 2024 for debugging).

National Center for Biotechnology Information (NCBI). Saccharomyces cerevisiae S288C Genome Assembly (GCF_000146045.2).
NCBI Datasets, https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000146045.2/. Accessed 4 Nov. 2024.

National Center for Biotechnology Information (NCBI). Sequence Read Archive (SRA) Run: ERR12205202.
NCBI SRA, https://trace.ncbi.nlm.nih.gov/Traces/?view=run_browser&acc=ERR12205202&display=download. Accessed 4 Nov. 2024.
}
