% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_calling.R
\name{variant_calling}
\alias{variant_calling}
\title{Calls Variants in a Filtered BAM File and Outputs VCF File}
\usage{
variant_calling(bamPath, refPath, outputVCF = "output_variants.vcf")
}
\arguments{
\item{outputVCF}{A character string for the output VCF file name. Default is "output_variants.vcf".}

\item{bamfilteredPath}{A character string specifying the path to the filtered BAM file.}

\item{referencePath}{A character string specifying the path to the reference genome in FASTA format.}
}
\value{
The path to the generated VCF file containing the identified variants.
}
\description{
This function takes a filtered BAM file and a reference genome in FASTA format to perform
variant calling. Detected variants, including SNPs, insertions, and deletions, are saved to a VCF file.
It utilizes VariantAnnotation and VariantTools packages for processing.
}
\examples{
# Example: Perform variant calling on a filtered BAM file
variantResult <- variant_calling(
  bamPath = "C:/Users/rjay1/Desktop/BCB410/LegitXMut/inst/extdata/filtered_output.bam",
  refPath = "C:/Users/rjay1/Desktop/BCB410/LegitXMut/inst/extdata/sequence_demo_ref.fasta",
  outputVCF = "C:/Users/rjay1/Desktop/BCB410/LegitXMut/inst/extdata/output_variants.vcf"
)

}
\references{
Obenchain V, Lawrence M, Carey V, Gogarten S, Shannon P, Morgan M (2014).
“VariantAnnotation: a Bioconductor package for exploration and annotation of
genetic variants.” Bioinformatics, 30(14), 2076-2078. doi:10.1093/bioinformatics/btu168.

Lawrence M, Degenhardt J, Gentleman R (2024). VariantTools: Tools for
Exploratory Analysis of Variant Calls. R package version 1.46.0.
}
