% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_FASTQ.R
\name{alignment_FASTQ}
\alias{alignment_FASTQ}
\title{Aligns FASTQ File to Reference Genome and Produces a BAM File}
\usage{
alignment_FASTQ(fastqPath, referencePath, outputBAM = "output.bam")
}
\arguments{
\item{fastqPath}{A character string for the path to the FASTQ file.}

\item{referencePath}{A character string for the path to the reference genome in FASTA format.}

\item{outputBAM}{A character string for the name of the output BAM file. Default is "output.bam".}
}
\value{
An object containing alignment information from the BAM file,
which can be used for downstream analysis, including variant calling and visualization.
}
\description{
A function to align a FASTQ file to a reference genome, generating a BAM file.
This function is suitable for handling large sequencing datasets and supports
alignment of long-read sequencing data, including Nanopore reads.
}
\examples{
# Example 1: Align a FASTQ file to a reference genome,remember to replace the file path to local while using.
alignmentResult <- alignment_FASTQ(
  fastqPath =  "C:/Users/rjay1/Desktop/BCB410/LegitXMut/data/sequence.fastq",
  referencePath =  "C:/Users/rjay1/Desktop/BCB410/LegitXMut/data/sequence.fastq",
  outputBAM =  "C:/Users/rjay1/Desktop/BCB410/LegitXMut/data/sequence.fastq"
)

}
\references{
Morgan M, Lawrence M, Aboyoun P, Gentleman R, and Pages H (2019).
“GenomicAlignments: Efficient alignments processing in R for NGS data.”
Bioconductor. doi:10.18129/B9.bioc.GenomicAlignments. Available at: https://bioconductor.org/packages/release/bioc/html/GenomicAlignments.html.

Liao Y, Smyth GK, Shi W (2019). “The R package Rsubread is easier, faster, cheaper
and better for alignment and quantification of RNA sequencing reads.”
Nucleic Acids Research, 47, e47. doi:10.1093/nar/gkz114.
Available at: https://academic.oup.com/nar/article/47/8/e47/5371636.

OpenAI. ChatGPT: Assistance with R function development for bioinformatics applications,
"Assignment". https://chat.openai.com (2023, accessed 30 October 2024).
}
