% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_FASTQ.R
\name{alignment_FASTQ}
\alias{alignment_FASTQ}
\title{Aligns FASTQ File to Reference Genome and Produces a BAM File, a VCF File and a Summary}
\usage{
alignment_FASTQ(
  fastqPath,
  referencePath,
  indels = 10,
  maxMismatches = 1000,
  outputBAM = "output.bam"
)
}
\arguments{
\item{fastqPath}{A character string for the path to the FASTQ file.}

\item{referencePath}{A character string for the path to the reference genome in FASTA format.}

\item{indels}{An integer specifying the maximum number of insertions or deletions allowed in the alignment. Default is 10.}

\item{maxMismatches}{An integer specifying the maximum number of mismatches allowed in the alignment. Default is 1000.}

\item{outputBAM}{A character string for the name of the output BAM file. Default is "output.bam".}
}
\value{
An object containing alignment information from the BAM file and information of variant calling from VCF file,
which can be used for downstream analysis, including variant calling and visualization.
}
\description{
A function to align a FASTQ file to a reference genome, generating a BAM file.
This function is suitable for handling large sequencing datasets and supports
alignment of Nanopore reads and quality control.
}
\details{
The LegitXMut package includes external demo data stored in the extdata directory,
and these external files are provided for demonstration purposes and support reproducibility of the analysis.
}
\examples{
# Example 1: Align a FASTQ file to a reference genome with custom parameters.
# Replace with actual test file path
fastqPath <- system.file("extdata", "ERR12205202.fastq", package = "LegitXMut")
referencePath <- system.file("extdata", "yeast.fna", package = "LegitXMut")
# Locate the extdata directory for outputs
outputDir <- system.file("extdata", package = "LegitXMut")
outputBAM <- file.path(outputDir, "aligned_output.bam")

mutateddemoalignment <- alignment_FASTQ(fastqPath =  fastqPath,
referencePath =  referencePath,
indels = 10,
maxMismatches = 1000,
outputBAM =  outputBAM)

}
\references{
Liao Y, Smyth GK, Shi W (2019)."The R package Rsubread is easier, faster, cheaper
and better for alignment and quantification of RNA sequencing reads."
Nucleic Acids Research, 47, e47. doi:10.1093/nar/gkz114.
Available at: https://academic.oup.com/nar/article/47/8/e47/5371636.

Morgan M, Lawrence M, Aboyoun P, Gentleman R, and Pages H (2019).
"GenomicAlignments: Efficient alignments processing in R for NGS data"
Bioconductor.doi:10.18129/B9.bioc.GenomicAlignments.Available at: https://bioconductor.org/packages/release/bioc/html/GenomicAlignments.html.

National Center for Biotechnology Information (NCBI). Saccharomyces cerevisiae S288C Genome Assembly (GCF_000146045.2).
NCBI Datasets, https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000146045.2/. Accessed 4 Nov. 2024.

National Center for Biotechnology Information (NCBI). Sequence Read Archive (SRA) Run: ERR12205202.
NCBI SRA, https://trace.ncbi.nlm.nih.gov/Traces/?view=run_browser&acc=ERR12205202&display=download. Accessed 4 Nov. 2024.

OpenAI. ChatGPT: Assistance with R function development for bioinformatics applications,
"Assignment".https://chat.openai.com (2023, accessed 5 November 2024 for debugging).
}
