% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLegitXMut.R
\name{runLegitXMut}
\alias{runLegitXMut}
\title{Launch Shiny App for LegitXMut}
\usage{
runLegitXMut()
}
\value{
No return value; launches a Shiny app in the browser.
}
\description{
A function that launches the Shiny app for the LegitXMut package.
The app provides a user-friendly interactive interface for alignment and visulization
using functions from the LegitXMut package.
}
\examples{
\dontrun{
LegitXMut::runLegitXMut()
}

}
\references{
Silva, Anjali. TestingPackage. GitHub, https://github.com/anjalisilva/TestingPackage. Accessed 5 Nov. 2024.
\href{https://shiny.rstudio.com/}{Link}
}
